move create_cmdline to schema to allow schema to handle setting of cmdline arguments
