+++ siliconcompiler/schema/schema_obj.py
@@ -12,6 +12,9 @@ import logging
 import os
 import re
 import pathlib
+import argparse
+import sys
+import shlex
 
 try:
     import yaml
@@ -1000,6 +1003,429 @@ class Schema:
                              'must be called on a complete keypath')
 
         return cfg['node']['default']['default']['value']
+
+    ###########################################################################
+    def create_cmdline(self,
+                       progname,
+                       description=None,
+                       switchlist=None,
+                       input_map=None,
+                       additional_args=None,
+                       version=None,
+                       print_banner=None,
+                       input_map_handler=None,
+                       preprocess_keys=None,
+                       post_process=None,
+                       logger=None):
+        """Creates an Schema command line interface.
+
+        Exposes parameters in the SC schema as command line switches,
+        simplifying creation of SC apps with a restricted set of schema
+        parameters exposed at the command line. The order of command
+        line switch settings parsed from the command line is as follows:
+
+         1. loglevel
+         2. fpga_partname
+         3. load_target('target')
+         4. read_manifest([cfg])
+         5. all other switches
+
+        The cmdline interface is implemented using the Python argparse package
+        and the following use restrictions apply.
+
+        * Help is accessed with the '-h' switch.
+        * Arguments that include spaces must be enclosed with double quotes.
+        * List parameters are entered individually. (ie. -y libdir1 -y libdir2)
+        * For parameters with Boolean types, the switch implies "true".
+        * Special characters (such as '-') must be enclosed in double quotes.
+        * Compiler compatible switches include: -D, -I, -O{0,1,2,3}
+        * Verilog legacy switch formats are supported: +libext+, +incdir+
+
+        Args:
+            progname (str): Name of program to be executed.
+            description (str): Short program description.
+            switchlist (list of str): List of SC parameter switches to expose
+                at the command line. By default all SC schema switches are
+                available. Parameter switches should be entered based on the
+                parameter 'switch' field in the schema. For parameters with
+                multiple switches, both will be accepted if any one is included
+                in this list.
+            input_map (dict of str): Dictionary mapping file extensions to input
+                filetypes. This is used to automatically assign positional
+                source arguments to ['input', 'fileset', ...] keypaths based on their file
+                extension. If None, the CLI will not accept positional source
+                arguments.
+            additional_args (dict of dict): Dictionary of extra arguments to add
+                to the command line parser, with the arguments matching the
+                argparse.add_argument() call.
+            version (str): Version to report when calling with -version
+            print_banner (function): Function callback to print command line banner
+            input_map_handler (function): Function callback handle inputs to the input map
+            preprocess_keys (function): Function callback to preprocess keys that need to be
+                corrected
+            post_process (function): Function callback to process arguments before returning
+
+        Returns:
+            None if additional_args is not provided, otherwise a dictionary with the
+                command line options detected from the additional_args
+
+        Examples:
+            >>> chip.create_cmdline(progname='sc-show',switchlist=['-input','-cfg'])
+            Creates a command line interface for 'sc-show' app.
+
+            >>> chip.create_cmdline(progname='sc', input_map={'v': ('rtl', 'verilog')})
+            All sources ending in .v will be stored in ['input', 'rtl', 'verilog']
+
+            >>> extra = chip.create_cmdline(progname='sc',
+                                            additional_args={'-demo': {'action': 'store_true'}})
+            Returns extra = {'demo': False/True}
+        """
+
+        if not logger:
+            logger = self.logger
+
+        # Argparse
+        parser = argparse.ArgumentParser(prog=progname,
+                                         prefix_chars='-+',
+                                         formatter_class=argparse.RawDescriptionHelpFormatter,
+                                         description=description,
+                                         allow_abbrev=False)
+
+        # Get a new schema, in case values have already been set
+        schema = Schema(logger=self.logger)
+
+        # Iterate over all keys from an empty schema to add parser arguments
+        used_switches = set()
+        for keypath in schema.allkeys():
+            # Fetch fields from leaf cell
+            helpstr = schema.get(*keypath, field='shorthelp')
+            typestr = schema.get(*keypath, field='type')
+            pernodestr = schema.get(*keypath, field='pernode')
+
+            # argparse 'dest' must be a string, so join keypath with commas
+            dest = '_'.join(keypath)
+
+            switchstrs, metavar = self._get_switches(schema, *keypath)
+
+            # Three switch types (bool, list, scalar)
+            if not switchlist or any(switch in switchlist for switch in switchstrs):
+                used_switches.update(switchstrs)
+                if typestr == 'bool':
+                    # Boolean type arguments
+                    if pernodestr == 'never':
+                        parser.add_argument(*switchstrs,
+                                            nargs='?',
+                                            metavar=metavar,
+                                            dest=dest,
+                                            const='true',
+                                            help=helpstr,
+                                            default=argparse.SUPPRESS)
+                    else:
+                        parser.add_argument(*switchstrs,
+                                            metavar=metavar,
+                                            nargs='?',
+                                            dest=dest,
+                                            action='append',
+                                            help=helpstr,
+                                            default=argparse.SUPPRESS)
+                elif re.match(r'\[', typestr) or pernodestr != 'never':
+                    # list type arguments
+                    parser.add_argument(*switchstrs,
+                                        metavar=metavar,
+                                        dest=dest,
+                                        action='append',
+                                        help=helpstr,
+                                        default=argparse.SUPPRESS)
+                else:
+                    # all the rest
+                    parser.add_argument(*switchstrs,
+                                        metavar=metavar,
+                                        dest=dest,
+                                        help=helpstr,
+                                        default=argparse.SUPPRESS)
+
+        # Check if there are invalid switches
+        if switchlist:
+            for switch in switchlist:
+                if switch not in used_switches:
+                    raise ValueError(f'{switch} is not a valid commandline argument')
+
+        if input_map is not None and input_map_handler:
+            parser.add_argument('source',
+                                nargs='*',
+                                help='Input files with filetype inferred by extension')
+
+        # Preprocess sys.argv to enable linux commandline switch formats
+        # (gcc, verilator, etc)
+        scargs = []
+
+        # Iterate from index 1, otherwise we end up with script name as a
+        # 'source' positional argument
+        for argument in sys.argv[1:]:
+            # Split switches with one character and a number after (O0,O1,O2)
+            opt = re.match(r'(\-\w)(\d+)', argument)
+            # Split assign switches (-DCFG_ASIC=1)
+            assign = re.search(r'(\-\w)(\w+\=\w+)', argument)
+            # Split plusargs (+incdir+/path)
+            plusarg = re.search(r'(\+\w+\+)(.*)', argument)
+            if opt:
+                scargs.append(opt.group(1))
+                scargs.append(opt.group(2))
+            elif plusarg:
+                scargs.append(plusarg.group(1))
+                scargs.append(plusarg.group(2))
+            elif assign:
+                scargs.append(assign.group(1))
+                scargs.append(assign.group(2))
+            else:
+                scargs.append(argument)
+
+        if version:
+            parser.add_argument('-version', action='version', version=version)
+
+        if additional_args:
+            # Add additional user specified arguments
+            arg_dests = []
+            for arg, arg_detail in additional_args.items():
+                argument = parser.add_argument(arg, **arg_detail)
+                arg_dests.append(argument.dest)
+            # rewrite additional_args with new dest information
+            additional_args = arg_dests
+
+        # Grab argument from pre-process sysargs
+        cmdargs = vars(parser.parse_args(scargs))
+
+        extra_params = None
+        if additional_args:
+            # Grab user specified arguments
+            extra_params = {}
+            for arg in additional_args:
+                if arg in cmdargs:
+                    extra_params[arg] = cmdargs[arg]
+                    # Remove from cmdargs
+                    del cmdargs[arg]
+
+        if print_banner:
+            print_banner()
+
+        # Set loglevel if set at command line
+        if 'option_loglevel' in cmdargs.keys():
+            log_level = cmdargs['option_loglevel']
+            if isinstance(log_level, list):
+                # if multiple found, pick the first one
+                log_level = log_level[0]
+            logger.setLevel(log_level.split()[-1])
+
+        # Read in all cfg files
+        if 'option_cfg' in cmdargs.keys():
+            for item in cmdargs['option_cfg']:
+                self.read_manifest(item, clobber=True, clear=True)
+
+        if input_map_handler:
+            # Map sources to ['input'] keypath.
+            if 'source' in cmdargs:
+                input_map_handler(cmdargs['source'])
+                # we don't want to handle this in the next loop
+                del cmdargs['source']
+
+        # Cycle through all command args and write to manifest
+        for dest, vals in cmdargs.items():
+            keypath = dest.split('_')
+
+            # Turn everything into a list for uniformity
+            if not isinstance(vals, list):
+                vals = [vals]
+
+            # Cycle through all items
+            for item in vals:
+                if item is None:
+                    # nargs=? leaves a None for booleans
+                    item = ''
+
+                if preprocess_keys:
+                    item = preprocess_keys(keypath, item)
+
+                num_free_keys = keypath.count('default')
+
+                switches, metavar = self._get_switches(schema, *keypath)
+                switchstr = '/'.join(switches)
+
+                if len(item.split(' ')) < num_free_keys + 1:
+                    # Error out if value provided doesn't have enough words to
+                    # fill in 'default' keys.
+                    raise ValueError(f'Invalid value {item} for switch {switchstr}. '
+                                     f'Expected format {metavar}.')
+
+                # We replace 'default' in keypath with first N words in provided
+                # value.
+                *free_keys, remainder = item.split(' ', num_free_keys)
+                args = [free_keys.pop(0) if key == 'default' else key for key in keypath]
+
+                # Remainder is the value we want to set, possibly with a step/index value beforehand
+                sctype = self.get(*keypath, field='type')
+                pernode = self.get(*keypath, field='pernode')
+                step, index = None, None
+                if pernode == 'required':
+                    try:
+                        step, index, val = remainder.split(' ', 2)
+                    except ValueError:
+                        self.logger.error(f"Invalid value '{item}' for switch {switchstr}. "
+                                          "Requires step and index before final value.")
+                elif pernode == 'optional':
+                    # Split on spaces, preserving items that are grouped in quotes
+                    items = shlex.split(remainder)
+                    if len(items) > 3:
+                        self.logger.error(f"Invalid value '{item}'' for switch {switchstr}. "
+                                          "Too many arguments, please wrap multiline "
+                                          "strings in quotes.")
+                        continue
+                    if sctype == 'bool':
+                        if len(items) == 3:
+                            step, index, val = items
+                        elif len(items) == 2:
+                            step, val = items
+                            if val != 'true' and val != 'false':
+                                index = val
+                                val = True
+                        elif len(items) == 1:
+                            val, = items
+                            if val != 'true' and val != 'false':
+                                step = val
+                                val = True
+                        else:
+                            val = True
+                    else:
+                        if len(items) == 3:
+                            step, index, val = items
+                        elif len(items) == 2:
+                            step, val = items
+                        else:
+                            val, = items
+                else:
+                    val = remainder
+
+                msg = f'Command line argument entered: {args} Value: {val}'
+                if step is not None:
+                    msg += f' Step: {step}'
+                if index is not None:
+                    msg += f' Index: {index}'
+                self.logger.info(msg)
+
+                # Storing in manifest
+                typestr = schema.get(*keypath, field='type')
+                if typestr.startswith('['):
+                    if self.valid(*args):
+                        self.add(*args, val, step=step, index=index)
+                    else:
+                        self.set(*args, val, step=step, index=index, clobber=True)
+                else:
+                    self.set(*args, val, step=step, index=index, clobber=True)
+
+        if post_process:
+            post_process(cmdargs)
+
+        return extra_params
+
+    ###########################################################################
+    def _get_switches(self, schema, *keypath):
+        '''Helper function for parsing switches and metavars for a keypath.'''
+        # Switch field fully describes switch format
+        switch = schema.get(*keypath, field='switch')
+
+        if switch is None:
+            switches = []
+        elif isinstance(switch, list):
+            switches = switch
+        else:
+            switches = [switch]
+        switchstrs = []
+
+        # parse out switch from metavar
+        # TODO: should we validate that metavar matches for each switch?
+        for switch in switches:
+            switchmatch = re.match(r'(-[\w_]+)\s+(.*)', switch)
+            gccmatch = re.match(r'(-[\w_]+)(.*)', switch)
+            plusmatch = re.match(r'(\+[\w_\+]+)(.*)', switch)
+
+            if switchmatch:
+                switchstr = switchmatch.group(1)
+                metavar = switchmatch.group(2)
+            elif gccmatch:
+                switchstr = gccmatch.group(1)
+                metavar = gccmatch.group(2)
+            elif plusmatch:
+                switchstr = plusmatch.group(1)
+                metavar = plusmatch.group(2)
+            switchstrs.append(switchstr)
+
+        return switchstrs, metavar
+
+    ###########################################################################
+    def read_manifest(self, filename, clear=True, clobber=True):
+        """
+        Reads a manifest from disk and merges it with the current manifest.
+
+        The file format read is determined by the filename suffix. Currently
+        json (*.json) and yaml(*.yaml) formats are supported.
+
+        Args:
+            filename (filepath): Path to a manifest file to be loaded.
+            clear (bool): If True, disables append operations for list type.
+            clobber (bool): If True, overwrites existing parameter value.
+
+        Examples:
+            >>> chip.read_manifest('mychip.json')
+            Loads the file mychip.json into the current Chip object.
+        """
+        schema = Schema(manifest=filename, logger=self.logger)
+
+        for keylist in schema.allkeys():
+            if keylist[0] in ('history', 'library'):
+                continue
+            if 'default' in keylist:
+                continue
+            typestr = schema.get(*keylist, field='type')
+            should_append = re.match(r'\[', typestr) and not clear
+
+            for val, step, index in schema._getvals(*keylist, return_defvalue=False):
+                # update value, handling scalars vs. lists
+                if should_append:
+                    self.add(*keylist, val, step=step, index=index)
+                else:
+                    self.set(*keylist, val, step=step, index=index, clobber=clobber)
+
+                # update other pernode fields
+                # TODO: only update these if clobber is successful
+                step_key = Schema.GLOBAL_KEY if not step else step
+                idx_key = Schema.GLOBAL_KEY if not index else index
+                for field in schema.getdict(*keylist)['node'][step_key][idx_key].keys():
+                    if field == 'value':
+                        continue
+                    v = schema.get(*keylist, step=step, index=index, field=field)
+                    if should_append:
+                        self.add(*keylist, v, step=step, index=index, field=field)
+                    else:
+                        self.set(*keylist, v, step=step, index=index, field=field)
+
+            # update other fields that a user might modify
+            for field in schema.getdict(*keylist).keys():
+                if field in ('node',):
+                    # skip these fields (node handled above)
+                    continue
+
+                # TODO: should we be taking into consideration clobber for these fields?
+                v = schema.get(*keylist, field=field)
+                self.set(*keylist, v, field=field)
+
+        # Read history, if we're not already reading into a job
+        if 'history' in schema.getkeys():
+            for historic_job in schema.getkeys('history'):
+                self.cfg['history'][historic_job] = schema.getdict('history', historic_job)
+
+        # TODO: better way to handle this?
+        if 'library' in schema.getkeys():
+            for libname in schema.getkeys('library'):
+                self.cfg['library'][libname] = schema.getdict('library', libname)
 
 
 if _has_yaml:
